created ast package - refactoring
