+++ de.variantsync.core/src/test/java/de/variantsync/core/JsonParserTest.java
@@ -8,6 +8,8 @@ import java.nio.file.Path;
 import java.util.Arrays;
 
 import de.variantsync.core.ast.AST;
+import de.variantsync.core.ast.JsonParserAST;
+import de.variantsync.core.ast.LineGrammar;
 import org.junit.Before;
 import org.junit.Test;
